% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_Delta}
\alias{change_Delta}
\title{A helper that takes a gurobi model object outputted from
dose_attributable_general or dose_thresh_attributable_one_sided
and changes the Delta parameter. Saves computation time by directly editing
the constraint involving Delta without having to reinitialize the other
constraints that are kept the same. Outputs a list analogous to output from
dose_attributable_general or dose_thresh_attributable_one_sided.}
\usage{
change_Delta(model, Delta, direction = "equal", TT)
}
\arguments{
\item{model}{A gurobi model object outputted from dose_attributable_general.}

\item{Delta}{The new Delta to test for.}

\item{direction}{The new direction to test}

\item{TT}{The observed test statistic.}
}
\value{
A gurobi model and solution.
}
\description{
A helper that takes a gurobi model object outputted from
dose_attributable_general or dose_thresh_attributable_one_sided
and changes the Delta parameter. Saves computation time by directly editing
the constraint involving Delta without having to reinitialize the other
constraints that are kept the same. Outputs a list analogous to output from
dose_attributable_general or dose_thresh_attributable_one_sided.
}
