% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE.R
\name{dose_attributable_general}
\alias{dose_attributable_general}
\title{Inference for general attributable effects in sensitivity analysis with
binary outcomes.}
\usage{
dose_attributable_general(
  Z,
  Q,
  index,
  gamma,
  alpha = 0.05,
  monotone = "increase",
  Delta,
  sign = "greater",
  trans = identity,
  alternative = "both",
  M = 10000,
  eps = 0.01,
  relax = "No",
  feasible = "Yes"
)
}
\arguments{
\item{Z}{A length N vector of (nonnegative) observed doses.}

\item{Q}{A length N vector of observed binary outcomes.}

\item{index}{A length N vector of indices indicating matched set membership.}

\item{gamma}{The nonnegative sensitivity parameter; gamma = 0 means no
unmeasured confounding.}

\item{alpha}{Level of the test.}

\item{monotone}{A string denoting what type of monotonicity to impose on the
potential outcomes; "increase", "decrease", or "none".}

\item{Delta}{A numeric for the attributable effect to be tested for.}

\item{sign}{A string taking value "greater", "less", or "equals" to denote the
constraint imposed on the attributable effect's relationship to Delta parameter.}

\item{trans}{The transformation of the doses to use for the test statistic.
Default is the identity function.}

\item{alternative}{A string; "lower", "upper", or "both" corresponding to what
deviates the sensitivity analysis non-rejection constraint should be on.}

\item{M}{The numeric penalty parameter to ensure correct sign relationship.}

\item{eps}{precision parameter for the objective function if solving feasible
= "Yes"}

\item{relax}{A string; "Yes" if solving the continuous relaxation; "No" if solving the
original MIQP formulation.}

\item{feasible}{A string; "Yes" if only want to see if the program is feasible -
meaning not rejecting, "No" if wanting the solve for the least rejectable allocation
of potential outcomes.}
}
\value{
A gurobi optimization object. If feasible is "Yes", an optimal solution found
means fail to reject. If feasible is "No", reject if the optimal value exceeds 0.
}
\description{
Inference for general attributable effects in sensitivity analysis with
binary outcomes.
}
