% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treat_out_match}
\alias{treat_out_match}
\title{Matched lead dataset}
\format{
\subsection{\code{treat_out_match}}{

A data frame with 4,134 rows and 4 columns:
\describe{
\item{treat}{The log of lead exposure level measured in micrograms per deciliter.}
\item{serious}{Whether the juvenile comitted a serious offense.}
\item{complain}{Whether the juvenile comitted an offense worthy of complaint.}
\item{match_ind}{Matched set membership.}
...
}
}
}
\source{
\url{https://scholarworks.iu.edu/dspace/handle/2022/25638}
}
\usage{
treat_out_match
}
\description{
A matched, trimmed dataset of early life lead exposure and
juvenile delinquency. There are 2007 matched sets.
}
\keyword{datasets}
