% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AE.R
\name{binary_thresh_attribut}
\alias{binary_thresh_attribut}
\title{Title}
\usage{
binary_thresh_attribut(Z, Q, index, gamma, thresh = 0, a = 1, trans = identity)
}
\arguments{
\item{Z}{A length N vector of (nonnegative) observed doses.}

\item{Q}{A length N vector of observed binary outcomes.}

\item{index}{A length N vector of indices indicating matched set membership.}

\item{gamma}{The nonnegative sensitivity parameter; gamma = 0 means no
unmeasured confounding.}

\item{thresh}{The dose threshold for the TAE}

\item{a}{The number of attributable effects to "test"}

\item{trans}{The transformation of the doses to use for the test statistic.
Default is the identity function.}
}
\value{
Either "reject" if not plausible/compatible, else a list containing a
p-value and dataframe of matched sets that have contribution to the test statistic
sorted in order of smallest mean reduction followed smallest variance reduction.
}
\description{
Title
}
