% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharp.R
\name{dose_sensitivity_mc_gen}
\alias{dose_sensitivity_mc_gen}
\title{Title}
\usage{
dose_sensitivity_mc_gen(
  Z,
  Q,
  index,
  mc,
  gamma,
  trans = identity,
  direct = "upper"
)
}
\arguments{
\item{Z}{A length N vector of (nonnegative) observed doses.}

\item{Q}{A length N vector of observed binary outcomes.}

\item{index}{A length N vector of indices indicating matched set membership.}

\item{mc}{An integer for the total number of Monte-Carlo samples.}

\item{gamma}{The nonnegative sensitivity parameter; gamma = 0 means no
unmeasured confounding.}

\item{trans}{The transformation of the doses to use for the test statistic.
Default is the identity function.}

\item{direct}{The direction of the test - "upper" or "lower"; default is upper.}
}
\value{
A length mc vector consisting of the Monte-Carlo samples of the
test statistic.
}
\description{
Title
}
